+++ flow_py_library_general_maths/array/assemble/addressable_to_int_list/test_addressable_to_int_list.py
@@ -1,7 +1,5 @@
-from pathlib import Path
-
 import numpy as np
-from flow.testing import FlowTest, flow_test
+from flow.test_framework import FlowTest, flow_test
 from flow_types import base
 
 from flow_py_library_general_maths.array.assemble.addressable_to_int_list.flow_addressable_to_int_list import (
@@ -11,8 +9,7 @@ from flow_py_library_general_maths.array.assemble.addressable_to_int_list.flow_a
 from flow_py_library_general_maths.util.utils_tests import basic_test_eval
 
 # Tests
-
-component_file = Path(__file__).parent
+component_file = "array_maths/assemble/addressable_to_int_list"
 
 
 def run_test_func(inputs, outputs, flow: FlowTest):