+++ flow_py_library_general_maths/array/sum/test_sum.py
@@ -1,15 +1,12 @@
-from pathlib import Path
-
 import numpy as np
-from flow.testing import FlowTest, flow_test
+from flow.test_framework import FlowTest, flow_test
 from flow_types import base
 
 from flow_py_library_general_maths.array.sum.flow_sum import inports, outports
 from flow_py_library_general_maths.util.utils_tests import basic_test_eval
 
 # Tests
-
-component_file = Path(__file__).parent
+component_file = "array_maths/sum"
 
 
 def run_test_func(inputs, outputs, flow: FlowTest):