+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.1.0 - 2020-07-2
+
+### Added
+
+- Assemble arrays and perform basic maths on them
+- Logging (instead of `print` statements)
+
+### Changed
+
+- Moved to new Python Abstraction Runtime: https://github.com/theengineeringco/flow-py/pull/92
+- Overhaul of component definitions (removed Runs and modified testing procedures)
+- Components no longer have a sub-function and a process, all maths is done directly in the process!
+
 ## v0.0.1 - 2020-07-02
 
 ### Added